+++ runtime/src/main/cpp/Porting.cpp
@@ -154,7 +154,7 @@ void consolePrintf(const char* format, ...) {
   va_start(args, format);
   int rv = vsnprintf_impl(buffer, sizeof(buffer), format, args);
   if (rv < 0) return; // TODO: this may be too much exotic, but should i try to print itoa(error) and terminate?
-  if (rv >= sizeof(buffer)) rv = sizeof(buffer) - 1;  // TODO: Consider realloc or report truncating.
+  if (static_cast<size_t>(rv) >= sizeof(buffer)) rv = sizeof(buffer) - 1;  // TODO: Consider realloc or report truncating.
   va_end(args);
   consoleWriteUtf8(buffer, rv);
 }
@@ -166,7 +166,7 @@ void consoleErrorf(const char* format, ...) {
   va_start(args, format);
   int rv = vsnprintf_impl(buffer, sizeof(buffer), format, args);
   if (rv < 0) return; // TODO: this may be too much exotic, but should i try to print itoa(error) and terminate?
-  if (rv >= sizeof(buffer)) rv = sizeof(buffer) - 1;  // TODO: Consider realloc or report truncating.
+  if (static_cast<size_t>(rv) >= sizeof(buffer)) rv = sizeof(buffer) - 1;  // TODO: Consider realloc or report truncating.
   va_end(args);
   consoleErrorUtf8(buffer, rv);
 }
@@ -451,7 +451,7 @@ extern "C" {
                 100000007UL,
                 1000000007UL
         };
-        int table_length = sizeof(primes)/sizeof(unsigned long);
+        size_t table_length = sizeof(primes)/sizeof(unsigned long);
 
         if (n > primes[table_length - 1]) konan::abort();
 
@@ -490,7 +490,7 @@ extern "C" {
 #ifdef KONAN_WASM
     // Some string.h functions.
     void *memcpy(void *dst, const void *src, size_t n) {
-        for (long i = 0; i != n; ++i)
+        for (size_t i = 0; i != n; ++i)
             *((char*)dst + i) = *((char*)src + i);
         return dst;
     }
@@ -507,7 +507,7 @@ extern "C" {
     }
 
     int memcmp(const void *s1, const void *s2, size_t n) {
-        for (long i = 0; i != n; ++i) {
+        for (size_t i = 0; i != n; ++i) {
             if (*((char*)s1 + i) != *((char*)s2 + i)) {
                 return *((char*)s1 + i) - *((char*)s2 + i);
             }
@@ -516,7 +516,7 @@ extern "C" {
     }
 
     void *memset(void *b, int c, size_t len) {
-        for (long i = 0; i != len; ++i) {
+        for (size_t i = 0; i != len; ++i) {
             *((char*)b + i) = c;
         }
         return b;
@@ -529,7 +529,7 @@ extern "C" {
     }
 
     size_t strnlen(const char *s, size_t maxlen) {
-        for (long i = 0; i<=maxlen; ++i) {
+        for (size_t i = 0; i<=maxlen; ++i) {
             if (s[i] == 0) return i;
         }
         return maxlen;
